backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
cryptography==42.0.0
PyJWT==2.8.0
Pillow==10.2.0
transformers==4.36.2
torch==2.1.2
spacy==3.7.2
PyPDF2==3.0.1
pdfplumber==0.10.3
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
gunicorn==21.2.0
prometheus-client==0.19.0

# LangChain and LangGraph for advanced RAG
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
langchain-core==0.1.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
